requirements.txt
pyspark==4.1.1
pandas==3.0.2
requests==2.33.1
python-dotenv==1.2.2
PyYAML==6.0.3
pyarrow==23.0.1
setuptools==82.0.1
